+++ src/main/java/kitchenpos/domain/Price.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain.product;
+package kitchenpos.domain;
 
 import java.math.BigDecimal;
 import java.util.List;
