+++ src/main/java/kitchenpos/domain/menu/MenuProduct.java
@@ -6,7 +6,7 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.ManyToOne;
 import javax.persistence.Transient;
-import kitchenpos.domain.product.Price;
+import kitchenpos.domain.Price;
 import kitchenpos.domain.product.Product;
 
 @Entity
