+++ src/main/java/kitchenpos/domain/menu/MenuProducts.java
@@ -8,7 +8,7 @@ import javax.persistence.CascadeType;
 import javax.persistence.Embeddable;
 import javax.persistence.FetchType;
 import javax.persistence.OneToMany;
-import kitchenpos.domain.product.Price;
+import kitchenpos.domain.Price;
 
 @Embeddable
 public class MenuProducts {
