+++ src/test/java/kitchenpos/domain/menu/MenuProductsTest.java
@@ -4,7 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import java.math.BigDecimal;
 import java.util.Arrays;
-import kitchenpos.domain.product.Price;
+import kitchenpos.domain.Price;
 import kitchenpos.domain.product.Product;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
