+++ src/test/java/kitchenpos/domain/product/ProductTest.java
@@ -3,6 +3,7 @@ package kitchenpos.domain.product;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.math.BigDecimal;
+import kitchenpos.domain.Price;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
